refactor(font): move the embedded font from the framework to the app
The web build needs a font because a browser exposes no font database and
both text stacks then draw nothing, silently. That font was baked into
matcha-ecs-widgets via include_bytes! behind cfg(web), which put an
app-level asset choice inside the framework.

It now belongs to matcha-web, and is registered in one call at startup:

    use matcha_ecs_widgets::WithDefaultFont;
    let app = UiEcs::new_async(..).await.with_default_font(FONT.clone());

New matcha-ecs-widgets/src/font.rs registers with *both* text stacks --
RichText/TextBox shape through parley, Text and Button labels through
suzuri -- because registering with only one leaves every button blank.
Individual RichText::font/TextBox::font remain for the "this one widget
uses a different font" case.

Notes worth keeping:

- The byte type is Arc<Vec<u8>>, not Arc<[u8]>: parley's Blob::new takes
  Arc<dyn AsRef<[u8]> + Send + Sync>, which needs a sized inner type, so
  Arc<[u8]> would force a full ~9.6 MB copy at registration.
- Deduplication is Arc::ptr_eq against kept-alive Arcs. A hash or a raw
  pointer key would let a dropped allocation's address be reused,
  silently skipping a registration and drawing nothing.
- Registration is not gated on the target: .font(..) is an explicit
  request, and honouring it only in a browser would be a platform
  difference with no signal. Native apps that never call it keep
  enumerating system fonts as before.
- It happens from after_spawn/patch, never from a RenderItem builder,
  because RichTextStyle::measure shapes through the same resource and so
  needs the font present before layout runs.
- UiEcs::world_mut exists so a widget crate can install its own startup
  resources without the core naming them.

The font bytes stay out of git (matcha-web/build.rs fetches and verifies
them on the first wasm build); the licence moves with the asset and stays
tracked, since built artifacts redistribute the font.
