+++ matcha-web/src/assets/NotoSansJP-VF.ttf.LICENSE
@@ -8,8 +8,7 @@ Canonical licence text: https://scripts.sil.org/OFL
 
 This font is redistributed here under that licence, which permits bundling and
 embedding in software (including compiled into a binary) provided this notice
-travels with it. The font is only compiled into web (wasm) builds; see
-`../embedded_font.rs` for why the web needs an embedded font at all.
+travels with it.
 
 The full licence text follows.
 
@@ -63,8 +62,7 @@ EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO ANY WARRANTIES OF
 MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT
 OF COPYRIGHT, PATENT, TRADEMARK, OR OTHER RIGHT. IN NO EVENT SHALL THE
 COPYRIGHT HOLDER BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-INCLUDING ANY GENERAL, SPECIAL, INDIRECT, INCIDENTAL, OR CONSEQUENTIAL
-DAMAGES, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 FROM, OUT OF THE USE OR INABILITY TO USE THE FONT SOFTWARE OR FROM
 OTHER DEALINGS IN THE FONT SOFTWARE.
